requirements.txt
pygame==2.5.2
python-chess==1.9.4
numpy==1.24.2
